Removed deprecated products
Removed Call API as it is a deprecated product. Call API is not part of the Voice API.
Added CLI as Caller Line Identity.
Removed Voice API from Sender ID. "Sender ID" is never used when referring to Voice products.
TTS API - Removed reference to Voice API as they are separate products and TTS API is not part of the Voice API.
Removed TTS Prompt API as it is a deprecated product. TTS Prompt API was not part of the Voice API.
Removed notation reference from Virtual Number as it does not make sense.
Voice - Removed VXML reference and corrected to JSON.
Voice API - Removed reference to TTS API, Call API and TTS-Prompt API as they are not and never have been part of the Voice API.
VXML - Removed reference as Nexmo does not support VXML.