[DEVX-2855] Fix spotlight form.
When we moved from jquery_ujs to rails-ujs, the form stopped working
because the browser started preventing the request due to `CORS policy`.
The reason why it started to fail is because of an upstream issue caused
by rails-ujs.
It is including the `x-csrf-token` header even if the domain is
different. Making the request server-side will prevent this issue from
happening again.